backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.2
pydantic-settings==2.1.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
python-docx==1.1.0
reportlab==4.0.8
httpx==0.25.2
beautifulsoup4==4.12.2
jinja2==3.1.2
